+++ src/QuickMarcEditor/utils.js
@@ -45,7 +45,7 @@ export const isFixedFieldRow = recordRow => recordRow.tag === '008';
 export const isMaterialCharsRecord = recordRow => recordRow.tag === '006';
 export const isPhysDescriptionRecord = recordRow => recordRow.tag === '007';
 
-export const getContentSubfieldValue = (content) => {
+export const getContentSubfieldValue = (content = '') => {
   return content.split(/\$/)
     .filter(str => str.length > 0)
     .reduce((acc, str) => {
@@ -362,7 +362,7 @@ export const getLocationValue = (value) => {
 };
 
 export const validateLocationSubfield = (field, locations) => {
-  const [, locationValue] = getLocationValue(field.content)?.split(' ');
+  const [, locationValue] = getLocationValue(field.content)?.replace(/\s+/, ' ').split(' ') || '';
 
   return !!locations.find(location => location.code === locationValue);
 };
@@ -402,7 +402,7 @@ export const checkDuplicate010Field = (marcRecords) => {
   const marc010Records = marcRecords.filter(({ tag }) => tag === '010');
 
   if (marc010Records.length > 1) {
-    return <FormattedMessage id="ui-quick-marc.record.error.010Field.multiple" />;
+    return <FormattedMessage id="ui-quick-marc.record.error.locControlNumber.multiple" />;
   }
 
   return undefined;
@@ -506,6 +506,12 @@ const validateMarcBibRecord = (marcRecords, linkableBibFields) => {
     return <FormattedMessage id="ui-quick-marc.record.error.title.multiple" />;
   }
 
+  const duplicate010FieldError = checkDuplicate010Field(marcRecords);
+
+  if (duplicate010FieldError) {
+    return duplicate010FieldError;
+  }
+
   const uncontrolledSubfields = ['uncontrolledAlpha', 'uncontrolledNumber'];
 
   const $9Error = validate$9InLinkable(marcRecords, linkableBibFields, uncontrolledSubfields);
@@ -538,6 +544,12 @@ const validateMarcHoldingsRecord = (marcRecords, locations) => {
 
   if (locationRecords.length > 1) {
     return <FormattedMessage id="ui-quick-marc.record.error.location.multiple" />;
+  }
+
+  if (locationRecords.length) {
+    if (getContentSubfieldValue(locationRecords[0].content).$b?.length > 1) {
+      return <FormattedMessage id="ui-quick-marc.record.error.field.onlyOneSubfield" values={{ fieldTag: '852', subField: '$b' }} />;
+    }
   }
 
   if (!validateLocationSubfield(marcRecords.find(({ tag }) => tag === '852'), locations)) {